requirements.txt
attrs==24.2.0
blinker==1.9.0
boto3==1.35.68
botocore==1.35.68
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
flasgger==0.9.7.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
gunicorn==23.0.0
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.0.2
packaging==24.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
s3transfer==0.10.4
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.32.1
Werkzeug==3.1.3
zipp==3.21.0
flask-cors
